VMXI Fix API data types
Change APi primitives to object types to allow nullable parameters and return values
Change primary keys to Long types
Add TemplateWrapper method to access specific return map field by key

+++ uk.ac.diamond.ispyb.api/src/uk/ac/diamond/ispyb/api/IspybDataCollectionApi.java
@@ -3,8 +3,8 @@ package uk.ac.diamond.ispyb.api;
 import java.io.Closeable;
 
 public interface IspybDataCollectionApi extends Closeable {
-	public int upsertDataCollectionMain(DataCollectionMain dataCollectionMain);
+	public Long upsertDataCollectionMain(DataCollectionMain dataCollectionMain);
 	public void updateDataCollectionExperiment(DataCollectionExperiment dataCollectionExperiment);
 	public void updateDataCollectionMachine(DataCollectionMachine dataCollectionMachine);
-	public int upsertDataCollectionGroup(DataCollectionGroup dataCollectionGroup);
+	public Long upsertDataCollectionGroup(DataCollectionGroup dataCollectionGroup);
 }
